fix(doc): remove unused TextDocumentItem struct feat(notification): introduce PublishDiagnosticsNotification struct fix(diagnostics): delete redundant workspaceDiagnostics.go file refactor(doc): remove unused TextDocumentItem struct fields
